Fix get Chat Names
