+++ .gitignore
@@ -8,3 +8,5 @@ test.py
 Searched \!.png
 .venv
 chromedriver
+Waiting.png
+test.py
