+++ Whatsapp/BaseWhatsapp.py
@@ -4,6 +4,7 @@ from selenium.webdriver.support.wait import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.common.keys import Keys
 from selenium.webdriver.common.action_chains import ActionChains
+from selenium.webdriver.common.actions.wheel_input import ScrollOrigin
 
 import json
 import os
@@ -25,12 +26,24 @@ class BaseWhatsapp:
         wait = WebDriverWait(self.browser, timeout)
         try:
             element = wait.until(
-                EC.presence_of_element_located((By.CLASS_NAME, cName)))
+                EC.presence_of_element_located((By.ID, cName)))
         except:
             exit("Element not found")
         return element
     
-    def __sendPageUP(self, count):
+    def __moveMouseToElement(self, element):
+        """Move mouse to an element"""
+        ActionChains(self.browser).move_to_element(element).perform()
+
+    def __mouseScroll(self, elem,count):
+        self.__moveMouseToElement(elem)
+        """Scroll mouse a specified number of times"""
+        origin = ScrollOrigin.from_element(elem)
+        ActionChains(self.browser).scroll_from_origin(origin, 0, count).perform()
+
+
+
+    def __sendPageUP(self,elem, count):
         """Send PAGE_UP key a specified number of times"""
         for i in range(count):
             ActionChains(self.browser).send_keys(Keys.PAGE_UP).perform()
