--- Whatsapp/xpaths.json
@@ -1,4 +1,0 @@
-{
-    "chatPannel":"/html/body/div[1]/div/div/div[3]/div/div[3]/div/div[3]",
-    "chatNames":"/html/body/div[1]/div/div/div[3]/div/div[3]/div/div[3]/div[1]/div/div/div/div/div/div/div/div/div/div/div/span"
-}
