+++ pyproject.toml
@@ -1,0 +1,7 @@
+[project]
+name = "whatsapp-selenium"
+version = "0.1.0"
+requires-python = ">=3.13"
+dependencies = [
+    "selenium==4.30.0",
+]
