+++ requirements.txt
@@ -1,1 +1,1 @@
-selenium==4.30.0
+selenium==4.35.0
